Refactor notification components: rename file
